Convert TranslateCustomTagDescriptor to TranslateCustomTagBlock.
