+++ common/lib/xmodule/setup.py
@@ -3,16 +3,12 @@
 from setuptools import find_packages, setup
 
 XMODULES = [
-    "book = xmodule.backcompat_module:TranslateCustomTagDescriptor",
-    "discuss = xmodule.backcompat_module:TranslateCustomTagDescriptor",
-    "image = xmodule.backcompat_module:TranslateCustomTagDescriptor",
     "section = xmodule.backcompat_module:SemanticSectionDescriptor",
-    "slides = xmodule.backcompat_module:TranslateCustomTagDescriptor",
-    "videodev = xmodule.backcompat_module:TranslateCustomTagDescriptor",
     "raw = xmodule.raw_module:RawDescriptor",
 ]
 XBLOCKS = [
     "about = xmodule.html_module:AboutBlock",
+    "book = xmodule.template_module:TranslateCustomTagBlock",
     "annotatable = xmodule.annotatable_module:AnnotatableBlock",
     "chapter = xmodule.seq_module:SectionBlock",
     "conditional = xmodule.conditional_module:ConditionalBlock",
@@ -20,9 +16,11 @@ XBLOCKS = [
     "course_info = xmodule.html_module:CourseInfoBlock",
     "customtag = xmodule.template_module:CustomTagBlock",
     "custom_tag_template = xmodule.template_module:CustomTagTemplateBlock",
+    "discuss = xmodule.template_module:TranslateCustomTagBlock",
     "error = xmodule.error_module:ErrorBlock",
     "hidden = xmodule.hidden_module:HiddenDescriptor",
     "html = xmodule.html_module:HtmlBlock",
+    "image = xmodule.template_module:TranslateCustomTagBlock",
     "library = xmodule.library_root_xblock:LibraryRoot",
     "library_content = xmodule.library_content_module:LibraryContentBlock",
     "library_sourced = xmodule.library_sourced_block:LibrarySourcedBlock",
@@ -33,12 +31,14 @@ XBLOCKS = [
     "problemset = xmodule.seq_module:SequenceBlock",
     "randomize = xmodule.randomize_module:RandomizeBlock",
     "sequential = xmodule.seq_module:SequenceBlock",
+    "slides = xmodule.template_module:TranslateCustomTagBlock",
     "split_test = xmodule.split_test_module:SplitTestBlock",
     "static_tab = xmodule.html_module:StaticTabBlock",
     "unit = xmodule.unit_block:UnitBlock",
     "vertical = xmodule.vertical_block:VerticalBlock",
     "video = xmodule.video_module:VideoBlock",
     "videoalpha = xmodule.video_module:VideoBlock",
+    "videodev = xmodule.template_module:TranslateCustomTagBlock",
     "videosequence = xmodule.seq_module:SequenceBlock",
     "word_cloud = xmodule.word_cloud_module:WordCloudBlock",
     "wrapper = xmodule.wrapper_module:WrapperBlock",
